remove all public setters from Style
do the used style stuff internally. Some used values can be resolved
inside of Style, the rest are stored on layout objects, and that's okay.

This is one step towards some kind of refactor to Style I'll do with
typescript. The prototype-based thing I was trying to do doesn't fit
well with types, and was probably a bit too weird anyways. I thought I
might grow to like it but I never did