refactor: turn level and bfc into traits
I wanted strong typing for these, but the old way meant that
the only way to represent `display: inline-block;` was by nesting
like: InlineBlockContainer -> BFC block container -> content but
this is not right. The specs imply that only one box should be
generated: BlockContainer -> content.

So if I wanted to continue using stronger types I'd need to
generate many different combinations of classes. That's when you
know inheritance isn't the right pattern.

Without the tests this is less code, and it matches some of the
WebKit code I've looked at.

The old way might have actually worked, however, floats
definitely need to be traits, and this generates a more minimal
tree.