correctly paint colored segments of arabic text
we can only call ctx.fillText if the boundaries are safe to break
on. if they're not, drop into fontkit and use canvas path APIs to
paint the smallest segments possible on the left and right side.
the idea of the algorithm is simple, but this took me days to get
right!

(canvas painter only; this is not possible with the HTML painter)