synchronous layout
this is just easier for people on the other side. I thought it
would also be faster to not have those microtask iteratons for
getFace, but amazingly performance is saying it's slower (but is
faster when I use the profiler?! I give up trying to understand).

the browser implementation (which doesn't exist yet) will be hurt
by this. my plan is to call hb.createBlob() and copy each font
passed to api.registerFont to it. that way it can be loaded
synchronously when needed, and we'll also be sharing memory between
harfbuzz and fontkit