perf: implement partial reshaping in line breaking
pretty big impact for large paragraphs: first 5,000 words of
The Little Prince total layout goes from 15ms to 8ms.

The Little Prince as-is (perf-2.ts, with divs) sees a modest
improvement of a few percent

I was reshaping the whole right side, which I don't think you even
need to necessarily do.

this is ideal now