perf: avoid calling next/prevGraphemeBreak
this was one of the dominators during shaping (not by much though,
the low hanging fruit is long gone). also the way I was putting
the segments together was contrived. I was super confused when I
first wrote that in the library 2 years ago

speeds up perf-2 by 13ms, finally getting below the 300ms barrier