perf: base the cascade/metrics cache on styles
this is too slow if you have a ton of tiny layouts, in which case
the font/metrics key takes long enough to calculate that it starts
showing up in profiles. I added hashMix to be used in the word
cache (which I haven't committed yet). when used for that, it
doesn't show up in the profiler very much (context: I'm testing
thousands of layouts for possible use in CellEngine).