remove metrics from paragraph
getFontMetrics is fast enough now that we can just call it on the
fly. I don't want to retain memory that's being calculated at the
start of layout anyways.

(it was better for when you have to shape, layout, layout, like
intrinsic size checks, but the cost of memory when layouts are
sitting around is not worth it. also this makes perf-2 run 2-3ms
faster)