perf: move glyph extraction to c++
it doesn't help node much, but it does seem to help Firefox pretty
consistently

I tried retaining the WASM memory instead of returning a copy and
freeing, but it was messy and would result in unnecessary retention
of the glyphs that weren't used during cascade. however, if I can
figure that out later, it might help Firefox more, which does spend
a decent amount of time copying to the return array.