refactor inline-level wrapping in block containers
the goal is to have enough context when a run is created to know
which IFC it will belong to so that I can eventually remove `text`
from the runs, and instead have them reference offsets into the
paragraph.

(I have no idea of it will really work or not, but this lead to
stronger type checking and shorter code. I'm trying to do that as
part of a rewrite of whitespace collapsing because it is slow and
cannot take into context inline-block right now, which is my real
initial goal).

so instead of wrapping boxes after a BFC is almost complete, wrap
them immediately when a BC is found in an inline.
`generateInlineBox` must yield after each box is created and resume
with context when called again