perf: rewrite whitespace collapsing
based on Firefox's whitespace "compression" in nsTextUtils. It
saddens me that this only took a day to get right, and I wrote the
Collapser in many confused trips to the library. It was needlessly
complicated, but turns out it was not that slow.

perf-2 improves by around 5ms and perf-1 is more consistently under
13ms

the real motivation is that I now have enough context to properly
(not) collapse whitespace when it is around an inline-block. That
was the main blocker, although I lost the code for inline-block.

paragraph text is now built during render tree construction, which
allows for removing the `text` property from Runs, improving memory.

I decided not to collapse whitespace on ShapedItems. Firefox
collapses the text pre-shaping, and doing so would involve
complicated glyph-walking, make the canvas backend have to do a lot
more work, and would potentially send too much work to HarfBuzz. I'm
not certain that it'd be any faster.