add some new planned properties
and change so that it's one ~feature per row rather than one
property.

min/max width/height are now planned
position: fixed is now planned

I got interested in fixed heh. I still think it's not needed in a
well organized app but this isn't the place to say that