block container's isFloat() can use style
rather than attributes

this also changes isInlineLevel to not depend on float. so it will
return true even if something floats. so now isInlineLevel, isFloat,
and future isAbsolute are "stupid" (don't check the other
properties that affect how it's really laid out). I'm doing this
because all of the layout logic has to check all of these anyways.
If I want them to be smart, can always do that in the future.