add logo to top of readme
I picked Raleway because it's flow-y and has some serif-y elements
that call out to the tradition of typography which CSS is always
drawing from. Text flowing into available space is like a natural
process, so I used a raindrop to connote nature. I originally used
"drop" because block boxes drop down, but I like the feeling of
nature better.