perf: precalculate used values on the Style
diverging it from computed styles even more

For the zoom property, which multiplies every length by a zoom
factor, I tested out continuing on the current path, which involved
a function for each of the affected properties. It was
significantly slower (perf-2 got 10ms slower pretty consistently).

I decided it's better to pre-calculate used values immediately.
These aren't fully used values per the spec: for example, a
percentage `width` will still be a percentage `width`, but they
are reduced as much as they can be. Since the style tree is far
smaller than the document tree, this is affordable.

I tried a hundred re-parses of the document in the demo site and
Style only took around 17ms. The difference in practice for
hyperscript users will be even more negligible.