support for multiple declared styles on an element
I started working on it because I was thinking about optimizing
changing the zoom property for a whole tree and I got totally
sidetracked. I don't even need this but it was interesting to
think about how to do it without affecting perf and it should
definitely make the API easier to use.