hello, Zig! Zig builds the C/C++ code and deps now
The build.sh was obviously bad because it referenced repos on my
computer without any versioning. CMake and Meson don't have any
kind of package management, so would require using submodules.

CMake and Meson are at least as hard to learn as the Zig build
system, but using the Zig build system enables using Zig, which
is a great fit for dropflow since there is so much generated code,
which could be addressed more elegantly with comptime.

Unfortunately the WASM is bigger, but perf-2 speeds up
significantly. I'm probably going to remove the language database
soon, so that will help reduce WASM size again.

Reflection: this took a lot of effort. Zig WASM targets don't
support a memory grow function (ziglang/zig#16429) so most of the
effort was bringing in binaryen and writing a bunch of code to
inject a function call next to memory.grow. Binaryen is needed
anyways for optimizations, but it took a while to get it to build.
I definitely would have been fine with emscripten and submodules
until I needed a reason to use Zig, but oh well, learned a lot.