new browser-based font api with custom descriptors
an old TODO. the way you had to pass a URL was weird, so now it
generates one (that's for logging). file:// urls load sync, and
no longer fail in Node (idk how that went uncaught so long, but
it has to do with the examples not using the public API... shame
on me).

Next up: update the examples and update loadNotoFonts

Ref #28
Also https://github.com/chearon/dropflow/issues/27#issuecomment-2587595432