formalize bit propagation up the tree
A refactor of box attributes, preprocess, and postprocess

Box attributes are now just "bits" since I use them for so many
things. There are two kinds right now: attributes and "has" bits.
The latter describe child content and propagate up the tree if
their parent meets certain critera.

The 4 Box "has" properties always did this, but I've now
formalized the concept so that IfcInline can use it neatly. While
that removed the extra tree-walking in IfcInline's ctor, it
somehow made performance a tiny bit worse (up to 0.5ms in perf-2).
It is a lot easier to think about so I'm willing to pay it.

Preprocess and postprocess are now known as prelayout and
postlayout and are stack-based instead of recursive. Sadly this
didn't help perf but I still would rather look at all that logic
in one place.

I split hasTextOrSizedInline into two since that was barely an
optimization and is clearer separated