move areas to Box
all Boxes can have padding/border/content. I had this on
BlockContainer because the Inline box model is stored as a
different rect type on the Paragraph. That was a mistake because
it has different paths for snapping, but I didn't have a solution
for breaking them across lines until I read about fragmentation.
I'll do it like the browsers do, probably Chrome, and store areas
on a fragmentation tree