intrinsics can now be determined before layout
I got curious about this when I started working on support for
images and it was awkward to hook into the existing system. I
looked at Firefox source and it seems they don't perform a full
layout either. The specs never talk about it as if it's a part of
layout. I think I was just scared to get this wrong, but most of
the details are already in createMarkIterator.

Cleans up the wholeItems/brokenItems discrepancy too.