improve, elaborate, and test hasForeground for inlines
I realized that because all of a style's paint properties will be
painted directly after each other, a single Style.hasPaint method
can be used to determine where that contributes: the foreground
or background.

It also helps convey that the background layer is not the same as
the background properties: it's simply a layer in which we paint.
Inlines paint backgrounds (and borders) in the foreground layer and
block containers paint backgrounds (and borders) in the background
layer.

Let's force each subclass of Box to implement this themselves,
since upcoming ImageBox will have different logic.