flow.load: include fonts that might be needed by struts
I should have known better than to try and do that. I spent a while
trying to keep the optimization: call flow.load on the layout tree,
but that requires preprocessing, post-order traversal, but that
required a parent stack, and looking at the children, but that was
unoptimal.

I checked browsers and they do make requests for the following:

<span style="font-family: f;"></span>

no padding, no border, no content. An unnecessary request, but
not worth optimizing.