refactor treeItems out; a few inline painting bugs
treeItems is basically all of the linebox items in one flat list,
in tree order. treeItems felt like a hack, and I'm trying to reduce
properties on Paragraph as both an optimization (this commit can
save a few MiB in CellEngine) and to prepare for the merging of
Paragraph with BlockContainer (of inlines).

In the process, I discovered that the paint calls for painting a
relatively positioned inline were both in the wrong order and
missing items. So this fixes a few bugs while being less code.

The solution took me days and days to figure out. Items are
already in logical order, but the inlines have to be walked too.
Involving lineboxes was my peril: painting purely from the inline
tree, background box list (which are also in logical order) and
item list works.

This function is difficult and complicated, though. One day, I'd
like to refactor visual order out into a mapping list that is
destroyed when layout is done, that way we can simply walk the
lineboxes, which would also be smaller in terms of memory.