part 1: add diverging BlockContainer types
This kind of sucks tbh. I wish that BlockContainer were defined as
BlockContainerOfInlines | BlockContainerOfBlocks, but the base
class is definitely needed. Having the union type for BlockLevel
and InlineLevel causes a ton of noise, but should maybe be
revisited. As-is, one could more easily forget to handle all types
of child box (isBlockContainerOfBlocks doesn't narrow away a
BlockContainer, for example).